requirements.txt
boto3>=1.34.0
botocore>=1.34.0
click>=8.1.7
python-dateutil>=2.8.2
tabulate>=0.9.0
python-dotenv>=1.0.0
rich>=13.7.0
